+++ src/re_com/core.cljs
@@ -2,37 +2,39 @@
   (:require-macros
    [re-com.core])
   (:require
-   [re-com.alert          :as alert]
-   [re-com.box            :as box]
-   [re-com.buttons        :as buttons]
-   [re-com.checkbox       :as checkbox]
-   [re-com.close-button   :as close-button]
-   [re-com.datepicker     :as datepicker]
-   [re-com.debug          :as debug]
-   [re-com.dropdown       :as dropdown]
-   [re-com.input-text     :as input-text]
-   [re-com.typeahead      :as typeahead]
-   [re-com.input-time     :as input-time]
-   [re-com.splits         :as splits]
-   [re-com.modal-panel    :as modal-panel]
-   [re-com.error-modal    :as error-modal]
-   [re-com.multi-select   :as multi-select]
-   [re-com.popover        :as popover]
-   [re-com.progress-bar   :as progress-bar]
-   [re-com.radio-button   :as radio-button]
-   [re-com.selection-list :as selection-list]
-   [re-com.slider         :as slider]
-   [re-com.tabs           :as tabs]
-   [re-com.tag-dropdown   :as tag-dropdown]
-   [re-com.text           :as text]
-   [re-com.throbber       :as throbber]
-   [re-com.tour           :as tour]
-   [re-com.tree-select    :as tree-select]
-   [re-com.v-table        :as v-table]
-   [re-com.simple-v-table :as simple-v-table]
-   [re-com.nested-grid    :as nested-grid]
-   [re-com.nested-v-grid  :as nested-v-grid]
-   [re-com.theme          :as theme]))
+   [re-com.alert           :as alert]
+   [re-com.box             :as box]
+   [re-com.buttons         :as buttons]
+   [re-com.checkbox        :as checkbox]
+   [re-com.close-button    :as close-button]
+   [re-com.datepicker      :as datepicker]
+   [re-com.debug           :as debug]
+   [re-com.dropdown        :as dropdown]
+   [re-com.input-text      :as input-text]
+   [re-com.typeahead       :as typeahead]
+   [re-com.input-time      :as input-time]
+   [re-com.splits          :as splits]
+   [re-com.modal-panel     :as modal-panel]
+   [re-com.error-modal     :as error-modal]
+   [re-com.multi-select    :as multi-select]
+   [re-com.popover         :as popover]
+   [re-com.progress-bar    :as progress-bar]
+   [re-com.radio-button    :as radio-button]
+   [re-com.selection-list  :as selection-list]
+   [re-com.slider          :as slider]
+   [re-com.horizontal-tabs :as horizontal-tabs]
+   [re-com.bar-tabs        :as bar-tabs]
+   [re-com.pill-tabs       :as pill-tabs]
+   [re-com.tag-dropdown    :as tag-dropdown]
+   [re-com.text            :as text]
+   [re-com.throbber        :as throbber]
+   [re-com.tour            :as tour]
+   [re-com.tree-select     :as tree-select]
+   [re-com.v-table         :as v-table]
+   [re-com.simple-v-table  :as simple-v-table]
+   [re-com.nested-grid     :as nested-grid]
+   [re-com.nested-v-grid   :as nested-v-grid]
+   [re-com.theme           :as theme]))
 
 ;; -----------------------------------------------------------------------------
 ;; re-com public API (see also re-com.util)
@@ -108,11 +110,13 @@
 (def tag-dropdown               tag-dropdown/tag-dropdown)
 (def text-tag                   tag-dropdown/text-tag)
 
-(def horizontal-tabs            tabs/horizontal-tabs)
-(def horizontal-bar-tabs        tabs/horizontal-bar-tabs)
-(def vertical-bar-tabs          tabs/vertical-bar-tabs)
-(def horizontal-pill-tabs       tabs/horizontal-pill-tabs)
-(def vertical-pill-tabs         tabs/vertical-pill-tabs)
+(def horizontal-tabs            horizontal-tabs/horizontal-tabs)
+(def bar-tabs                   bar-tabs/bar-tabs)
+(def horizontal-bar-tabs        bar-tabs/horizontal-bar-tabs)
+(def vertical-bar-tabs          bar-tabs/vertical-bar-tabs)
+(def pill-tabs                  pill-tabs/pill-tabs)
+(def horizontal-pill-tabs       pill-tabs/horizontal-pill-tabs)
+(def vertical-pill-tabs         pill-tabs/vertical-pill-tabs)
 
 (def label                      text/label)
 (def p                          text/p)