+++ src/re_com/debug.cljs
@@ -82,6 +82,8 @@
         src
         (assoc :data-rc-src (str file ":" line))))))
 
+
+
 (defn component-stack
   ([el]
    (component-stack [] el))