+++ src/re_com/theme.cljs
@@ -3,10 +3,12 @@
   (:require
    [re-com.theme.util :as tu]
    [re-com.part :as part]
+   [re-com.debug :as debug]
    [re-com.theme.default :as theme.default]))
 
 (def ^:dynamic *variables* theme.default/variables)
 (def ^:dynamic *pre-user* theme.default/main)
+(def ^:dynamic *bootstrap* theme.default/bootstrap)
 (def ^:dynamic *base* theme.default/base)
 (def ^:dynamic *main* theme.default/main)
 (def ^:dynamic *user* nil)
@@ -28,7 +30,9 @@
 (def merge-props tu/merge-props)
 
 (defn re-com-meta [{:keys [part] :as props}]
-  (update props :class merge-class (part/css-class* part)))
+  (-> props
+      (update :class merge-class (part/css-class* part))
+     #_ (merge (debug/->attr props))))
 
 (defn comp [component-local-pre-theme component-local-theme]
   (clojure.core/apply
@@ -37,6 +41,7 @@
                   re-com-meta
                   *user*
                   *main*
+                  *bootstrap*
                   *base*
                   component-local-pre-theme
                   *pre-user*