+++ src/re_com/theme/default.cljs
@@ -72,3 +72,7 @@
 (defmethod main :default [props] props)
 
 (defmethod base :default [props] props)
+
+(defmulti bootstrap :part)
+
+(defmethod bootstrap :default [props] props)